// src/types.ts
export interface SRCProps {
  children: any
  //   width
  width: string
  w_414: string
  w_600: string
  w_768: string
  w_992: string
  w_1200: string
  w_1440: string
  w_1920: string
  //   height
  height: string
  h_414: string
  h_600: string
  h_768: string
  h_992: string
  h_1200: string
  h_1440: string
  h_1920: string
  //   display
  display: string
  d_414: string
  d_600: string
  d_768: string
  d_992: string
  d_1200: string
  d_1440: string
  d_1920: string
  //   flexDirection
  flexDirection: string
  fd_414: string
  fd_600: string
  fd_768: string
  fd_992: string
  fd_1200: string
  fd_1440: string
  fd_1920: string
  //   alignItems
  alignItems: string
  alignItems_414: string
  alignItems_600: string
  alignItems_768: string
  alignItems_992: string
  alignItems_1200: string
  alignItems_1440: string
  alignItems_1920: string
  //   margin
  margin: string
  m_414: string
  m_600: string
  m_768: string
  m_992: string
  m_1200: string
  m_1440: string
  m_1920: string
  //   marginTop
  mt: string
  mt_414: string
  mt_600: string
  mt_768: string
  mt_992: string
  mt_1200: string
  mt_1440: string
  mt_1920: string
  //   marginLeft
  ml: string
  ml_414: string
  ml_600: string
  ml_768: string
  ml_992: string
  ml_1200: string
  ml_1440: string
  ml_1920: string
  //   marginRight
  mr: string
  mr_414: string
  mr_600: string
  mr_768: string
  mr_992: string
  mr_1200: string
  mr_1440: string
  mr_1920: string
  //   marginBottom
  mb: string
  mb_414: string
  mb_600: string
  mb_768: string
  mb_992: string
  mb_1200: string
  mb_1440: string
  mb_1920: string
  //   padding
  padding: string
  p_414: string
  p_600: string
  p_768: string
  p_992: string
  p_1200: string
  p_1440: string
  p_1920: string
  //   paddingTop
  pt: string
  pt_414: string
  pt_600: string
  pt_768: string
  pt_992: string
  pt_1200: string
  pt_1440: string
  pt_1920: string
  //   paddingLeft
  pl: string
  pl_414: string
  pl_600: string
  pl_768: string
  pl_992: string
  pl_1200: string
  pl_1440: string
  pl_1920: string
  //   paddingRight
  pr: string
  pr_414: string
  pr_600: string
  pr_768: string
  pr_992: string
  pr_1200: string
  pr_1440: string
  pr_1920: string
  //   paddingBottom
  pb: string
  pb_414: string
  pb_600: string
  pb_768: string
  pb_992: string
  pb_1200: string
  pb_1440: string
  pb_1920: string
}